requirements.txt
pandas
pytest-xdist
snowflake-ml-python
